+++ zerver/actions/create_user.py
@@ -16,6 +16,7 @@ from zerver.actions.invites import notify_invites_changed
 from zerver.actions.message_send import internal_send_private_message, internal_send_stream_message
 from zerver.actions.streams import bulk_add_subscriptions, send_peer_subscriber_events
 from zerver.actions.user_groups import do_send_user_group_members_update_event
+from zerver.actions.user_settings import check_change_full_name
 from zerver.actions.users import change_user_is_active, get_service_dicts_for_bot
 from zerver.lib.avatar import avatar_url
 from zerver.lib.create_user import create_user
@@ -551,6 +552,12 @@ def do_reactivate_user(user_profile: UserProfile, *, acting_user: Optional[UserP
                 }
             ).decode(),
         )
+
+        if user_profile.full_name.endswith(" (spammer)"):
+            check_change_full_name(
+                user_profile, user_profile.full_name.replace(" (spammer)", ""), acting_user
+            )
+
         do_increment_logging_stat(
             user_profile.realm,
             COUNT_STATS["active_users_log:is_bot:day"],